requirements/base.txt
aiogram
Babel
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-rosetta
django-filter
django-modeltranslation
django-redis
django_recaptcha
cryptography==41.0.7
celery==5.3.6
redis==5.0.1
flower==2.0.0



